# server/docs/swagger/swagger.yaml
basePath: /api
definitions:
  http.AssetHistoryDTO:
    properties:
      cash:
        type: number
      date:
        type: string
      loan:
        type: number
      other:
        type: number
      remark:
        type: string
      stockValue:
        type: number
    type: object
  http.CodeSourceDTO:
    properties:
      cacheKey:
        type: string
      code:
        type: string
      codeKey:
        type: string
      defaultOpen:
        type: boolean
      description:
        type: string
      editorHeight:
        type: integer
      lang:
        type: string
      pageKey:
        type: string
      theme:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      zoomable:
        type: boolean
    type: object
  http.CodeSourceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.CodeSourceDTO'
        type: array
    type: object
  http.CodeSourceVersionDTO:
    properties:
      changeNote:
        type: string
      code:
        type: string
      createdAt:
        type: string
      version:
        type: integer
    type: object
  http.ConsumedLegRequest:
    properties:
      consumedShares:
        type: number
      legId:
        type: integer
    required:
    - legId
    type: object
  http.CreateAssetHistoryRequest:
    properties:
      cash:
        type: number
      date:
        type: string
      loan:
        type: number
      other:
        type: number
      remark:
        type: string
      stockValue:
        type: number
    required:
    - date
    type: object
  http.CreateDailyLogRequest:
    properties:
      consumedLegs:
        description: ConsumedLegs 是本次提交消耗掉的历史未匹配腿，后端据此扣减 stock_unmatched_legs。
        items:
          $ref: '#/definitions/http.ConsumedLegRequest'
        type: array
      date:
        type: string
      marker:
        description: '"！"=当日有操作 / ""=无操作'
        type: string
      monthlySwingRevenue:
        description: MonthlySwingRevenue 是本月波段收益的最新累计净收益（由前端基于历史值+当日增量计算后传入）。
        type: number
      monthlyTRevenue:
        description: MonthlyTRevenue 是本月做T的最新累计净收益（由前端基于历史值+当日增量计算后传入）。
        type: number
      newUnmatchedLegs:
        description: NewUnmatchedLegs 是当日新产生、仍未匹配完的腿，会持久化等待未来撮合。
        items:
          $ref: '#/definitions/http.UnmatchedLegRequest'
        type: array
      positions:
        description: Positions 是当日收盘后的最新持仓成本和数量。
        items:
          $ref: '#/definitions/http.PositionUpdateRequest'
        type: array
      review:
        $ref: '#/definitions/http.ReviewRequest'
      swingRecords:
        description: SwingRecords 是当日跨日撮合成功的波段收益明细，撮合计算在前端完成后提交。
        items:
          $ref: '#/definitions/http.SwingRecordRequest'
        type: array
      tRecords:
        items:
          $ref: '#/definitions/http.TRecordRequest'
        type: array
      trades:
        items:
          $ref: '#/definitions/http.TradeRequest'
        type: array
    required:
    - date
    type: object
  http.DashboardResponse:
    properties:
      assetHistory:
        items:
          $ref: '#/definitions/http.AssetHistoryDTO'
        type: array
      monthly:
        items:
          $ref: '#/definitions/http.MonthlyDTO'
        type: array
      otherIncome:
        items:
          $ref: '#/definitions/http.OtherIncomeDTO'
        type: array
      positions:
        items:
          $ref: '#/definitions/http.PositionDTO'
        type: array
    type: object
  http.MonthlyDTO:
    properties:
      month:
        type: string
      remark:
        type: string
      swingRevenue:
        type: number
      tRevenue:
        type: number
      tTarget:
        type: number
    type: object
  http.OtherIncomeDTO:
    properties:
      amount:
        type: number
      date:
        type: string
      desc:
        type: string
      remark:
        type: string
    type: object
  http.PositionBaseDTO:
    properties:
      cost:
        type: number
      date:
        type: string
      remark:
        type: string
      shares:
        type: number
    type: object
  http.PositionDTO:
    properties:
      base:
        $ref: '#/definitions/http.PositionBaseDTO'
      code:
        type: string
      cost:
        type: number
      remark:
        type: string
      shares:
        type: number
      stock:
        type: string
      target:
        items:
          $ref: '#/definitions/http.PositionTargetDTO'
        type: array
    type: object
  http.PositionSnapshotAsOfDTO:
    properties:
      date:
        type: string
      positions:
        items:
          $ref: '#/definitions/http.PositionSnapshotDTO'
        type: array
    type: object
  http.PositionSnapshotDTO:
    properties:
      code:
        type: string
      cost:
        type: number
      price:
        type: number
      shares:
        type: number
      stock:
        type: string
    type: object
  http.PositionSnapshotDatesDTO:
    additionalProperties:
      items:
        $ref: '#/definitions/http.PositionSnapshotDTO'
      type: array
    type: object
  http.PositionTargetDTO:
    properties:
      cost:
        type: number
      date:
        type: string
      price:
        type: number
      remark:
        type: string
      shares:
        type: number
      status:
        type: string
    type: object
  http.PositionUpdateRequest:
    properties:
      code:
        type: string
      cost:
        type: number
      price:
        type: number
      shares:
        type: number
      stock:
        type: string
    required:
    - code
    type: object
  http.ReviewRequest:
    properties:
      feeling:
        type: string
      market:
        type: string
      nextPlan:
        type: string
    type: object
  http.SwingRecordRequest:
    properties:
      buyDate:
        type: string
      desc:
        type: string
      fee:
        type: number
      grossProfit:
        type: number
      netRevenue:
        type: number
      sellDate:
        type: string
      stock:
        type: string
      tax:
        type: number
    type: object
  http.TRecordRequest:
    properties:
      desc:
        type: string
      fee:
        type: number
      grossProfit:
        type: number
      netRevenue:
        type: number
      stock:
        type: string
      tax:
        type: number
    type: object
  http.TradeRequest:
    properties:
      action:
        type: string
      code:
        type: string
      hasIssue:
        type: boolean
      price:
        type: number
      shares:
        type: number
      stock:
        type: string
    type: object
  http.UnmatchedLegDTO:
    properties:
      action:
        type: string
      code:
        type: string
      date:
        type: string
      id:
        type: integer
      price:
        type: number
      remainingShares:
        type: number
      stock:
        type: string
      totalFee:
        type: number
      totalShares:
        type: number
    type: object
  http.UnmatchedLegRequest:
    properties:
      action:
        type: string
      code:
        type: string
      price:
        type: number
      remainingShares:
        type: number
      stock:
        type: string
      totalFee:
        type: number
      totalShares:
        type: number
    type: object
  httpx.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8888
info:
  contact: {}
  description: 个人知识库后端服务，提供股票仪表盘和代码内容管理接口。
  title: Interview Server API
  version: "1.0"
paths:
  /admin/stock/asset-history:
    post:
      consumes:
      - application/json
      parameters:
      - description: 资产快照数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateAssetHistoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.AssetHistoryDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 新增资产快照
      tags:
      - stock
  /admin/stock/daily-log:
    post:
      consumes:
      - application/json
      parameters:
      - description: 每日收盘数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateDailyLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 提交每日收盘记录
      tags:
      - stock
  /code-sources/pages/{pageKey}:
    get:
      description: 返回指定 pageKey 下的所有公开记录，前端可用来一次性预加载整页数据。
      parameters:
      - description: 页面/业务域 key，如 expand-mermaid
        in: path
        name: pageKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CodeSourceListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 批量获取页面代码内容
      tags:
      - code-sources
  /code-sources/pages/{pageKey}/{codeKey}:
    get:
      description: 按 pageKey + codeKey 精确查询，找不到返回 404。前端 useCodeSource hook 对接此接口。
      parameters:
      - description: 页面/业务域 key，如 summary-editor
        in: path
        name: pageKey
        required: true
        type: string
      - description: 内容 key，如 demo-mermaid
        in: path
        name: codeKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CodeSourceDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取代码内容
      tags:
      - code-sources
  /code-sources/pages/{pageKey}/{codeKey}/versions:
    get:
      description: 返回指定 pageKey + codeKey 记录的历史版本列表，按版本号倒序排列。
      parameters:
      - description: 页面/业务域 key
        in: path
        name: pageKey
        required: true
        type: string
      - description: 内容 key
        in: path
        name: codeKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.CodeSourceVersionDTO'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取代码内容编辑历史
      tags:
      - code-sources
  /stock/asset-history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.AssetHistoryDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取资产历史曲线
      tags:
      - stock
  /stock/dashboard:
    get:
      description: 一次返回资产历史、持仓、月度收益、其他收入，前端 StockDashboard 直接消费。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DashboardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取股票仪表盘聚合数据
      tags:
      - stock
  /stock/monthly:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.MonthlyDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取月度 T 收益
      tags:
      - stock
  /stock/other-income:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.OtherIncomeDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取其他收入/支出
      tags:
      - stock
  /stock/position-snapshot-as-of/{date}:
    get:
      parameters:
      - description: YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PositionSnapshotAsOfDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取某天或之前最近一次的持仓快照
      tags:
      - stock
  /stock/position-snapshots:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PositionSnapshotDatesDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取所有有持仓快照的日期明细
      tags:
      - stock
  /stock/positions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.PositionDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取当前持仓
      tags:
      - stock
  /stock/unmatched-positions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.UnmatchedLegDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: 获取未匹配做T仓位
      tags:
      - stock
schemes:
- http
- https
swagger: "2.0"